services/analytics/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
gunicorn>=21.2,<23.0

pydantic-settings>=2.2,<3.0

httpx>=0.26,<1.0
pandas>=2.0,<3.0
yfinance>=0.2,<1.0

PyJWT>=2.8,<3.0